% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cost_function}
\alias{cost_function}
\title{Cost function for optimization.}
\usage{
cost_function(y, rsm_output, nfact, xnam)
}
\arguments{
\item{y}{Factors to optimize.}

\item{rsm_output}{Second order model fit.}

\item{nfact}{Number of factor values.}

\item{xnam}{Formula names for \code{\link{rsm}}.}
}
\description{
\code{cost_function}
}
\keyword{internal}
