% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{desirability}
\alias{desirability}
\title{Quadratic desirability function.}
\usage{
desirability(
  factors,
  Target,
  w,
  a,
  b,
  c,
  nfact,
  xnam,
  tuner_rsm,
  Pred_min,
  Pred_max
)
}
\arguments{
\item{factors}{Factors (tuning variables).}

\item{Target}{Target value for the response.}

\item{a}{Parameter of the quadratic desirabilty function, defaults to a = 0.}

\item{b}{Parameter of the quadratic desirabilty function, defaults to b = 0.2.}

\item{c}{Parameter of the quadratic desirabilty function, defaults to c = 0.8.}

\item{nfact}{Number of factor values.}

\item{xnam}{Formula names for \code{\link{rsm}}.}

\item{tuner_rsm}{Second order model fit.}

\item{Pred_min}{Minimum value which can be reached by the model within the factor ranges.}

\item{Pred_max}{Maximum value which can be reached by the model within the factor ranges.}
}
\description{
\code{desirability} is the cost function used for optimization.
}
\keyword{internal}
