% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{desirability_meas}
\alias{desirability_meas}
\title{Overall desirability function for the measured responses.}
\usage{
desirability_meas(y, Target, w, Pred_min, Pred_max)
}
\arguments{
\item{y}{Array of the measured responses (typically resolution or sensitivity).}

\item{Target}{Vector of the target values for each response.}

\item{w}{Vector of weights for each response.}

\item{Pred_min}{Vector of minimum values which can be reached by the model
within the factor ranges for each response variable.}

\item{Pred_max}{Vector of maximum value which can be reached by the model
within the factor ranges for each response variable.}
}
\description{
\code{desirability_meas} is the overall desirability function for the 
measured responses.
}
\details{
Note that the desirability of the measured responses also depends on the
rsm model (via Pred_min and Pred_max).
}
\keyword{internal}
