% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{desirability_overall}
\alias{desirability_overall}
\title{Overall desirability function.}
\usage{
desirability_overall(
  factors,
  Target,
  w,
  nfact,
  xnam,
  tuner_rsm,
  Pred_min,
  Pred_max
)
}
\arguments{
\item{factors}{Factors (tuning variables).}

\item{Target}{Vector of the target values for each response.}

\item{w}{Vector of weights for each the response.}

\item{nfact}{Number of factor values.}

\item{xnam}{Formula names for \code{\link{rsm}}.}

\item{tuner_rsm}{Second order model fit.}

\item{Pred_min}{Vector of minimum values which can be reached by the model
within the factor ranges for each response variable.}

\item{Pred_max}{Vector of maximum value which can be reached by the model
within the factor ranges for each response variable.}
}
\description{
\code{desirability_overall} is the cost function for optimization with more
than one response variable. It is the weighted mean of the individual
desirability functions.
}
\keyword{internal}
