% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{integrand}
\alias{integrand}
\title{Integrand for tof period calculation.}
\usage{
integrand(x, E, L, V, H)
}
\arguments{
\item{x}{Axial distance from the end electrode.}

\item{E}{Potential energy at the turning point (normalized with the mean 
energy (K/K_0).}

\item{L}{Vector of electrode lengths normalized with H ("lenght"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}

\item{H}{Height of the mirror electrodes (the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{integrand} is the integrand for the time-of-flight period calculation.
}
\keyword{internal}
