% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_coeffs}
\alias{plot_coeffs}
\title{Plots the fit coefficients of the response surface model.}
\usage{
plot_coeffs(tuner_rsm, ylab = "", factors)
}
\arguments{
\item{tuner_rsm}{Second order model fit.}

\item{ylab}{y-axis label.}

\item{factors}{Factors.}
}
\description{
\code{plot_coeffs} generates a plot with the fit coefficients of the response
surface model with 95% confidence interval.
}
\keyword{internal}
