% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{potential}
\alias{potential}
\title{Calculates the axial potential of gridless planar mirrors.}
\usage{
potential(x, L, V, H)
}
\arguments{
\item{x}{Axial distance from the end electrode.}

\item{L}{Vector of electrode lengths normalized with H ("lenght"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}

\item{H}{Height of the mirror electrodes (the same for all electrodes).}
}
\value{
Axial potential at \code{x}.
}
\description{
\code{potential} calculates the axial potential of gridless planar mirrors.
}
\references{
Yavor, M.I. et al. (2018), High performance gridless ion mirrors 
for multi-reflection time-of-flight and electrostatic trap mass analyzers, 
\emph{International Journal of Mass Spectrometry}, \strong{426},
1-11, doi:10.1016/j.ijms.2018.01.009.
}
\keyword{internal}
