% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{potential_inv}
\alias{potential_inv}
\title{Calculates the inverse of the axial potential function.}
\usage{
potential_inv(y, L, V, H)
}
\arguments{
\item{y}{Axial potential.}

\item{L}{Vector of electrode lengths normalized with H ("lenght"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}

\item{H}{Height of the mirror electrodes (the same for all electrodes).}
}
\value{
Position x where the axial potential = y.
}
\description{
\code{potential_inv} calculates the position x given the axial potential (inverse
of axial potential function).
}
\details{
The inverse of the axial potential function is used to calculate the turning 
point inside the mirror for a given kinetic energy.
}
\keyword{internal}
