% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMIONtuneR-package.R
\docType{package}
\name{SIMIONtuneR}
\alias{SIMIONtuneR}
\title{SIMION Parameter Tuning using Response Surface Methodology.}
\description{
The R package provides an interface to the ion optics simulation program
SIMION for parameter optmization using design of expriments (DoE) and
response surface methodology (RSM).
}
\author{
Patrick Sturm <sturm@tofwerk.com>
}
