% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tofperiod}
\alias{tofperiod}
\title{Time-of-flight period calculation.}
\usage{
tofperiod(E, x1, L, V, H)
}
\arguments{
\item{E}{Vector of potential energies at the turning point (normalized with
the mean energy (K/K_0).}

\item{x1}{End distance of the particle, typically the time-of-flight focal 
point of the mirror.}

\item{L}{Vector of electrode lengths normalized with H ("lenght"/\code{H}).}

\item{V}{Vector of voltages normalized with the mean energy ("voltage"/K_0)}

\item{H}{Height of the mirror electrodes (the same for all electrodes).}
}
\value{
time-of-flight.
}
\description{
\code{tofperiod} calculates the time-of-flight period inside the mirror.
}
\details{
This integrates the particle motion from the turning point x0 to the distance
x1. Because here we are only interested in relative time-of-flight deviations
all constant factors (e.g. \code{2*sqrt(2*m*amu/e)}) are omitted and the 
time-of-flight is in arbitrary units.
}
\keyword{internal}
