% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_results}
\alias{plot_results}
\title{Plots results.}
\usage{
plot_results(
  pltly,
  k,
  n_steps,
  resultdir,
  result,
  runs,
  run_optimized,
  result_optimized,
  bestpoint_predicted,
  bestpoint_selected,
  xylabel
)
}
\arguments{
\item{pltly}{Plotly object.}

\item{k}{Step number.}

\item{n_steps}{Total number of steps.}

\item{resultdir}{Results directory (used for legend text).}

\item{result}{Results (Resolving power and sensitivity).}

\item{runs}{Box-Behnken design data.}

\item{run_optimized}{Optimized bestpoint control values.}

\item{result_optimized}{Verified optimized bestpoint result.}

\item{bestpoint_predicted}{Model prediction of bestpoint result.}

\item{bestpoint_selected}{Selected bestpoint.}

\item{xylabel}{Axis labels.}
}
\description{
\code{plot_results} plots results.
}
\keyword{internal}
