% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_results}
\alias{plot_results}
\title{Plots results.}
\usage{
plot_results(
  pltly,
  k,
  n_steps,
  resultdir,
  result,
  runs,
  bestpoint_run,
  bestpoint_result,
  bestpoint_predicted,
  xylabel
)
}
\arguments{
\item{pltly}{Plotly object.}

\item{k}{Step number.}

\item{n_steps}{Total number of steps.}

\item{resultdir}{Results directory (used for legend text).}

\item{result}{Results (Resolving power and sensitivity).}

\item{runs}{Box-Behnken design data.}

\item{bestpoint_run}{Bestpoint control values.}

\item{bestpoint_result}{Verified bestpoint result.}

\item{bestpoint_predicted}{Model prediction of bestpoint result.}

\item{xylabel}{Axis labels.}
}
\description{
\code{plot_results} plots results.
}
\keyword{internal}
