% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMIONtuneR.R
\name{run_GLPMtuneR}
\alias{run_GLPMtuneR}
\title{Runs gridless planar mirror optimization.}
\usage{
run_GLPMtuneR(tuneR_config)
}
\arguments{
\item{tuneR_config}{path and name of tuneR_config file (.toml)}
}
\description{
\code{run_GLPMtuneR} runs gridless planar mirror optimization.
}
\details{
The experiment needs to be configured in a tuneR_config file (TOML file format).
If you want to start from previous best point, copy bestpoint_run.txt to tuneR_dir
(delete bestpoint.txt if you want to start with starting values from tuneR_config).
}
\examples{
tuneR_config = system.file("GLPMtuneR_config.toml", package = "SIMIONtuneR")
run_GLPMtuneR(tuneR_config)

}
