% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMIONtuneR.R
\name{run_SIMIONtuneR}
\alias{run_SIMIONtuneR}
\title{Runs SIMIONtuneR simulation.}
\usage{
run_SIMIONtuneR(tuneR_config, nogui = TRUE)
}
\arguments{
\item{tuneR_config}{path and name of tuneR_config file (.toml)}

\item{nogui}{run SIMION with --nogui option (\code{TRUE} (default) or \code{FALSE})}
}
\description{
\code{run_SIMIONtuneR} runs SIMIONtuneR simulation.
}
\details{
The experiment needs to be configured in a tuneR_config file (TOML file format).
See \code{system.file("SIMIONtuneR_config.toml", package = "SIMIONtuneR")} for 
a template.

If you want to start from a previous best point, copy bestpoint_run.txt to 
the tuneR directory (delete bestpoint_run.txt if you want to start with 
starting values from tuneR_config).
}
\examples{
tuneR_config = system.file("SIMIONtuneR_config.toml", package = "SIMIONtuneR")
run_SIMIONtuner(tuneR_config)

}
