% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMIONtuneR.R
\name{run_SIMIONtuneR}
\alias{run_SIMIONtuneR}
\title{Runs SIMIONtuneR simulation.}
\usage{
run_SIMIONtuneR(tuneR_config, nogui = TRUE)
}
\arguments{
\item{tuneR_config}{path and name of tuneR_config file (.toml)}

\item{nogui}{run SIMION with --nogui option (\code{TRUE} (default) or \code{FALSE})}
}
\description{
\code{run_SIMIONtuneR} runs SIMIONtuneR simulation.
}
\details{
The experiment needs to be configured in a tuneR_config file (TOML file format).
If you want to start from previous best point, copy bestpoint.txt to tuneR_dir
(delete bestpoint.txt if you want to start with starting values from tuneR_config).
}
\examples{
configfile = system.file("tuneR_config.toml", package = "SIMIONtuner")
run_SIMIONtuner(configfile)

}
