% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMIONtuneR-package.R
\docType{package}
\name{SIMIONtuneR}
\alias{SIMIONtuneR}
\alias{SIMIONtuneR-package}
\title{SIMION Parameter Tuning using Response-Surface Design.}
\description{
The package provides functions for optmization in SIMION software using 
design-of-expriments methods.
}
\author{
Patrick Sturm <sturm@tofwerk.com>
}
